Licenseportalapplication/requirements.txt
asgiref==3.2.3
Babel==2.8.0
Django==3.0.4
djangorestframework==3.12.2
python-dateutil==2.8.1
pytz==2019.3
six==1.15.0
sqlparse==0.3.1
